+++ completion/git-machete.completion.bash
@@ -21,7 +21,7 @@ _git_machete() {
     local reapply_opts="-f --fork-point="
     local slide_out_opts="-d --down-fork-point= -M --merge -n --no-edit-merge --no-interactive-rebase"
     local squash_opts="-f --fork-point="
-    local status_opts="--color= -L --list-commits-with-hashes -l --list-commits"
+    local status_opts="--color= -L --list-commits-with-hashes -l --list-commits -M --merge"
     local traverse_opts="-F --fetch -l --list-commits -M --merge -n --no-edit-merge --no-interactive-rebase --return-to= --start-from= -w --whole -W -y --yes"
     local update_opts="-f --fork-point= -M --merge -n --no-edit-merge --no-interactive-rebase"