+++ RELEASE_NOTES.md
@@ -5,7 +5,7 @@
 - improved: slide-out can target branches with any number of downstream (child) branches
 - fixed: detection of no-op rebase cases in fork-point algorithm
 - fixed: if a branch is merged to its parent, `git machete status -l` now always displays an empty list of commits
-- imprpved: `traverse` and `status` can detect rebase and squash merges via diff-tree, use `--merge` to fallback to strict merge detection
+- improved: `traverse` and `status` can detect rebase and squash merges via diff-tree, use `--merge` to fallback to strict merge detection
 
 ## New in git-machete 3.1.1